Add linter unconvert and fix issues
Removes unnecessary type conversions

* unnecessary conversion

+++ internal/cmd/cli/gitcloner/cloner.go
@@ -174,7 +174,7 @@ func createAuthFromOpts(opts *GitCloner) (transport.AuthMethod, error) {
 			return nil, fmt.Errorf("failed to read GitHub app private key from file: %w", err)
 		}
 
-		auth, err := appAuthGetter.Get(int64(opts.GitHubAppID), int64(opts.GitHubAppInstallation), key)
+		auth, err := appAuthGetter.Get(opts.GitHubAppID, opts.GitHubAppInstallation, key)
 		if err != nil {
 			return nil, err
 		}
